Final Masterthesis Code/requirements.txt
lxml==5.3.0
networkx==3.4.2
numpy==1.24.4
pandas==2.2.3
psutil==5.9.8
python_dateutil==2.8.2
